test: move skills-cli integration coverage to e2e lane
